requirements.txt
scipy
matplotlib
ftfy
torch
torchvision
nltk
pytorch-lightning
transformers
autofaiss
wandb
parlai